going production

+++ app/api/board/route.js
@@ -25,6 +25,13 @@ export async function POST(req) {
 
     const user = await User.findById(session.user.id);
 
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "You need to have a subscription to create a board" },
+        { status: 403 }
+      );
+    }
+
     const board = await Board.create({
       userID: user._id,
       name: body.name,
@@ -59,11 +66,21 @@ export async function DELETE(req) {
       return NextResponse.json({ error: "Not authorized" }, { status: 401 });
     }
 
+    const user = await User.findById(session?.user?.id);
+
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "You need to have a subscription to create a board" },
+        { status: 403 }
+      );
+    }
+
     await Board.deleteOne({ _id: boardId, userID: session?.user?.id });
 
-    const user = await User.findById(session?.user?.id);
     user.boards = user.boards.filter((id) => id.toString() !== boardId);
+
     await user.save();
+
     return NextResponse.json({});
   } catch (e) {
     return NextResponse.json({ error: e.message }, { status: 500 });
